+++ src/KristofferStrube.Blazor.FileSystemAccess/FileSystemDirectoryHandle.cs
@@ -9,10 +9,14 @@ public class FileSystemDirectoryHandle : FileSystemHandle
 {
 
     public static new FileSystemDirectoryHandle Create(IJSRuntime jSRuntime, IJSObjectReference jSReference)
-        => Create(jSRuntime, jSReference, FileSystemAccessOptions.DefaultInstance);
+    {
+        return Create(jSRuntime, jSReference, FileSystemAccessOptions.DefaultInstance);
+    }
 
     public static new FileSystemDirectoryHandle Create(IJSRuntime jSRuntime, IJSObjectReference jSReference, FileSystemAccessOptions options)
-        => new(jSRuntime, jSReference, options);
+    {
+        return new(jSRuntime, jSReference, options);
+    }
 
     internal FileSystemDirectoryHandle(IJSRuntime jSRuntime, IJSObjectReference jSReference, FileSystemAccessOptions options) : base(jSRuntime, jSReference, options) { }